rename WorkflowExec to workflow-exec
