+++ packages/frinx-workflow-ui/src/App.js
@@ -13,7 +13,7 @@ import buildReducer from './store/reducers/builder';
 import bulkReducer from './store/reducers/bulk';
 import searchReducer from './store/reducers/searchExecs';
 import WorkflowDefinitions from './pages/workflow-list/WorkflowDefs/workflow-definitions';
-import WorkflowExec from './pages/workflow-list/WorkflowExec/workflow-exec';
+import WorkflowExec from './pages/workflow-list/workflow-exec/workflow-exec';
 import Scheduling from './pages/workflow-list/Scheduling/Scheduling';
 import EventListeners from './pages/workflow-list/EventListeners/event-listeners';
 import TaskList from './pages/workflow-list/Tasks/task-list';
