+++ packages/frinx-workflow-ui/src/pages/workflow-list/workflow-exec/DetailsModal/DetailsModal.css
