+++ packages/frinx-workflow-ui/src/pages/workflow-list/workflow-exec/DetailsModal/details-modal.tsx
